Aktueller Stand vor to_fen-Vergleich

+++ src/types.hpp
@@ -78,7 +78,7 @@ struct GameState{
   CastlingRight castling_rights;
   Square ep_square;
   uint16_t half_move_clock;
-  uint8_t total_move_count;
+  uint16_t total_move_count;
 
   //Nicht im FEN
   Piece captured_piece;
